+++ src/components/image/Image.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import { ImageProps } from './types'
+import { Image as ImageBase } from 'react-native'
+
+const Image: React.FC<ImageProps> = ({ src, style }) => {
+    return <ImageBase source={src} style={style} />
+}
+
+export default Image