+++ src/components/image/types.ts
@@ -1,0 +1,6 @@
+import { ImageSourcePropType, ImageStyle } from 'react-native'
+
+export interface ImageProps {
+    src: ImageSourcePropType
+    style?: ImageStyle | any
+}