+++ src/components/index.ts
@@ -4,5 +4,6 @@ import PasswordField from './form/password-field'
 import Text from './text'
 import Title from './title'
 import TextField from './form/text-field'
+import Image from './image'
 
-export { Button, Icon, PasswordField, Text, TextField, Title }
+export { Button, Icon, PasswordField, Text, TextField, Title, Image }